requirements.txt
streamlit
pandas
TCSP
snownlp
scikit-learn
matplotlib
datetime
pytz
openpyxl
plotly.figure_factory